+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Snap, create: Fixed possible deadlock
+
 ## [1.28.0] - 2024-07-11
 
 ### Added